Expire OAuth access tokens
Can see any use-case, currently, that requires not expireing tokens. The use-cases, login through OAuth, a new access token is generated if not there (or expired).
Unkown if in use, but could cause an issue with any app that comunicate with API, but then best to use refresh access tokens and have apps refresh token.