Upgrade Gem dependenices
Changed to a supported version of AWS S3, which seems to only be requred by Paperclip.
Also changed a aws config file to use the new lib, but unclear where this file is used.
Fixed a lower version of sdoc, where a lower depenency psych was causing issues.
Upgrade doorkeeper to the latest version.