Fix a few broken active admin listing pages
Since ransack is limiting the columns allowed,
and Activeadmin does not pass a auth object,
decided to manualy set the allowed filter fields,
based on what was already passed with out auth.
There is also a custom concern, that if the context
of the ransack attributes allowed is activeadmin,
treat as if was superadmin, since only they allowed.